core: up max request size to 10MB
Increase the max request size from 1MB to 10MB. It's pretty common for
people to run into this. Requests can grow unexpectedly large from:

* using batching to submit multiple transactions at once
* spend actions being satisfied using many small outputs
* large reference data objects or large asset definitions
* our JSON wire format with hex-encoded raw transactions

When testing locally, I hit the 1MB limit signing a single transaction
with ~750 issuances, ~750 outputs and a small amount of json metadata
on each issuance (in the asset definition).